requirements.txt
flask
flask-restful
flask-migrate
flask-sqlalchemy
marshmallow >= 3
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
gunicorn
psycopg2
pytest
pytest-factoryboy
pytest-runner
pytest-flask
pytest-celery
factory_boy
scalpl
